refactor: 루틴으로 예약하는 로직을 RoutineService로 이동 (#187)
